requirements.txt
opencv-python
torch
torchvision
ultralytics
numpy
scipy
diffusers
transformers
accelerate
xformers
pillow
